+++ coresim/main.cpp
@@ -30,13 +30,13 @@ std::priority_queue<Event*, std::vector<Event*>, EventComparator> event_queue;
 std::deque<Flow*> flows_to_schedule;
 std::deque<Event*> flow_arrivals;
 
-uint32_t num_outstanding_packets = 0;
-uint32_t max_outstanding_packets = 0;
-uint32_t num_outstanding_packets_at_50 = 0;
-uint32_t num_outstanding_packets_at_100 = 0;
-uint32_t arrival_packets_at_50 = 0;
-uint32_t arrival_packets_at_100 = 0;
-uint32_t arrival_packets_count = 0;
+long long num_outstanding_packets = 0;
+long long max_outstanding_packets = 0;
+long long num_outstanding_packets_at_50 = 0;
+long long num_outstanding_packets_at_100 = 0;
+long long arrival_packets_at_50 = 0;
+long long arrival_packets_at_100 = 0;
+long long arrival_packets_count = 0;
 uint32_t total_finished_flows = 0;
 uint32_t duplicated_packets_received = 0;
